Incoming chat messages queued up for broadcast
Each chat message received needs to find the game the player is a part
of, if any. This is a bit time consuming because there isn't a quick way
to get to it. Perhaps a HashMap<PlayerID, Player> would be better.

Messages are queued up as pending for each game slot. Not yet
implemented is the Server broadcasting this message for everyone in that
game slot.