Fixed Client/Server infinite message spam
The client was getting an UpdateReply but with last_acked_msg_seq_num as
None. The check failed so it remained None. Later in the chat message
loop, the 'continue' test never passed so we kept sending more messages
via Update to the client. Ad infinitum.

My testing here shows it's okay, but I wouldn't be surprised if there
are more bugs. Unit Testing will help a lot with this.