[WIP] Assign header field ID's for struct/enums
Every struct member and variant's fields must have a header field
identifier if it will be decoded. This is now built up by examining the
parsed AST instead of hardcoding it.

Parsing and header field allocation takes place when Wireshark calls
into to register the library instead of in `lazy_static!{}`. The
`lazy_static` instantiation order doesn't appear to be defined (at
least I can't find it), so I had issues where calling `get()` on the
alloctor took place prior to any `register()` calls. Mutexes
were added to allow the modification of static globals (as needed by
the library) at run-time in order to accomplish this.