[WIP] Working more on decoding buffer data, see notes
- Added `byteorder` to help with byte casting
- Added the decoding of Enumerator's fields
- Split `Sizing::Variable` so that it can differentiate between a `Vector`
and an `Option`. Previously there wasn't a difference, and this
presented an issue when looking at how many bytes to consume after the
`bincode` marker in the buffer. Options are encoded by one byte (0 or 1)
to represent `Some` or `None`. Vectors use 8 bytes for size. It was
impossible to determine how many bytes to consume due to this `bincode`
design discrepency.