Fixed bug where EttInfo addresses were moving after registration
The issue was simultaneously updating the `EttInfo::ett_items` and
`EttInfo::addresses` (which points to the former) at the same time.
After so many items were pushed into `ett_items`, the vector was resized
and moved location, but the `addresses` were not updated so they now
pointed to something else. The wireshark dissector caught this b/c the
addresses array that it dereferences was not "-1" by value. The solution
is to build up the `ett_items` array first, and then after registration
create the `addresses` list that points to it.