Bug fix: Dissector now parses all netwayste rust source files
The addition of `PingPong` as a field in several `Packet` variants broke
the dissector because it was only looking for definitions in `net.rs`
and not any other folder in the `netwayste` crate. After parsing the
`net.rs` file, the `ett` had no record of a `PingPong` because it is
defined in a separate file. This patch allows the dissector to parse all
rust source files within the netwayste directory to broaden the scope of
where data types can exist. Think of it as prep for when `net.rs` is
refactored into more managable components.