[WIP] Escape logic moved to layer when no keypress handled
It is my belief that key handlers should return `Handled` when the enact
upon a key-press. A bug has been fixed when this wasn't the case where
the Escape key would not only release focus of a textfield, but also pop
the screen stack. That is not the desired behavior. Handlers like mouse
move should still return `NotHandled` from where I stand now.

Two bugs still remain:

1) Changing state to the `server_list` ignores any Escape keys
and does not pop back to the main menu.

2) Logic still exists within the `client.rs` to handle the first
key-press. Case in point, when the Return key is pushed to enter focus
into the chatbox. In escape's case, this returns "unhandled key code"
because we don't... handle it :P. Need to look into moving some of this
stuff to the layer if possible. The escape is processed once the game
area gains focus. So maybe after running immediately gain focus in the
game area? Seems like a hack.