More bug fixes for gamearea mouse and key support
- Textfield focus now works correctly using mouse buttons and
keyboard keys to either gain or lose focus.

- Added a new `EventType` to allow a widget to request focus of another
widget. The gamearea widget needs to request focus of the chatbox's
textfield on the Enter key. Since these widgets have no tree
relationship (the only belong to the same layer), the request-for-focus
is needed.

- Refactored some existing code into a `switch_focus()` routine.

- Gamearea is now the default focus for the `Run` screen so that all
keyboard and mouse events are properly forwarded to the widget.

- Finally removed `process_running_inputs()`, and the associated
conditional flags as they are no longer required.