Fixed game-area panning regression bug
The gamearea's keypress handler updates the `game_area_state`. A
snapshot of the game area's state is taken prior to processing new UI
events. It needs to be updated after all events have been processed. I
considered moving this logic earlier in `update()`, but decided against
it; there will always be a need to clean-up, and it doesn't really
matter if you do it at the end or beginning of this routine. Special
circumstances may state otherwise but I don't think this is one of
those.