notebook successfully working for client connect and server response
The client port was hardcoded so that an endpoint wrapper could be
created for it and used in the response by the 'server'. This will need
to be swapped out for pulling the port from the bounded socket at
runtime.

The notebook will still sometimes fail on the assertion that the notify
lists are non-empty when getting them, but rerunning these cells acts as
a workaround. Adding in a `sleep(n)` call does not resolve the issue.